requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
email-validator==2.1.1
pydantic-settings==2.2.1
httpx==0.27.0
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
pytest==8.1.1
python-dotenv==1.0.1
openai==1.14.0
gunicorn==21.2.0
uvicorn[standard]==0.28.0
